Refactor package installation classes into separate files
Separated existing installation package classes (WordPressPackage, PluginPackage, ThemePackage) from `package.js` into their own files(`WpPackage.js`, `PluginPackage.js`, `ThemePackage.js`). This refactoring improves code organization, clarity, and separation of concerns. The classes are now directly exported from their files, making it easier to import them separately where necessary.
