+++ lib/Package.js
@@ -3,17 +3,15 @@ const path = require('path');
 const { exec } = require('child_process');
 const {
   getDownloadUrl,
-  getWordPressDownloadUrl,
   installNpmPackages,
-  replaceDbConstant,
-  replaceEmptySalts, installComposer, replaceDbConstantBool
+  installComposer
 } = require('./utils');
 
 const {
   downloadFile,
   extractZip,
   renameFolder
-} = require('./fs');
+} = require('./utils/fs');
 
 class Package {
   /**
@@ -25,7 +23,7 @@ class Package {
    * @param {Object} paths - the object containing the paths
    * @param {string} paths.rootFolder - the root folder of the application
    * @param {string} paths.tempDir - the temporary directory
-   * @param {string} paths.baseFolder - the temporary directory
+   * @param {string} paths.baseFolder - the script base directory
    * @param {string} paths.destFolder - The destination folder for package installation.
    */
   constructor (config, paths) {
@@ -81,7 +79,6 @@ class Package {
    *
    * @param {string} packageUrl - The URL of the package to download.
    * @param {string} packageName - The name of the package.
-   * @param {string} targetDirectory - The directory where the package will be saved.
    * @return {Promise} A promise that resolves when the package is successfully downloaded and saved, or rejects with an error.
    */
   async downloadPackage (packageUrl, packageName) {
@@ -134,116 +131,4 @@ class Package {
   }
 }
 
-class WordPressPackage extends Package {
-  /**
-   * Installs WordPress with the specified version and language.
-   *
-   * @param {string} version - The version of WordPress to install.
-   * @param {string} language - The language of WordPress to install.
-   */
-  async installWordPress (version, language) {
-    const downloadUrl = getWordPressDownloadUrl(version, language);
-
-    try {
-      const destinationPath = path.join(this.rootFolder, this.config.name);
-
-      if (fs.existsSync(destinationPath)) {
-        console.log('🔄️ WordPress folder already exists. Skipping download.');
-      } else {
-        // Download WordPress
-        return await this.execDownload(`wordpress-${version}.zip`, downloadUrl).then(() => {
-          // Copy WordPress folder to destination path
-          renameFolder(path.join(this.tempDir, 'wordpress'), destinationPath);
-          console.log(`🆗 WordPress installed successfully in ${destinationPath}`);
-        });
-      }
-    } catch (error) {
-      console.error('🔴 Error downloading or installing WordPress:', error);
-    }
-  }
-
-  /**
-   * Sets up the WordPress configuration by copying the sample config file,
-   * replacing the placeholder values with the actual configuration values,
-   * and saving the updated config file.
-   *
-   * @return {void} This function does not return anything.
-   */
-  async setupWordPressConfig () {
-    const configPath = path.join(this.baseFolder, 'wp-config.php');
-
-    try {
-      if (fs.existsSync(configPath)) {
-        console.log('🆗 WordPress configuration already set up. updating...');
-      } else {
-        const sampleConfigPath = path.join(this.baseFolder, 'wp-config-sample.php');
-        // Copy wp-config-sample.php to wp-config.php
-        fs.copyFileSync(sampleConfigPath, configPath);
-      }
-
-      // Read the content of wp-config.php
-      let configContent = fs.readFileSync(configPath, 'utf8');
-
-      // Update database name, username, password, and other settings based on user-defined config
-      configContent = replaceDbConstant(configContent, 'DB_NAME', this.config.wordpress.config.DB_NAME);
-      configContent = replaceDbConstant(configContent, 'DB_USER', this.config.wordpress.config.DB_USER);
-      configContent = replaceDbConstant(configContent, 'DB_PASSWORD', this.config.wordpress.config.DB_PASSWORD);
-      configContent = replaceDbConstant(configContent, 'DB_HOST', this.config.wordpress.config.DB_HOST);
-      configContent = replaceDbConstant(configContent, 'DB_CHARSET', this.config.wordpress.config.DB_CHARSET);
-
-      configContent = replaceDbConstantBool(configContent, 'WP_DEBUG', this.config.wordpress.config.WP_DEBUG);
-
-      configContent = replaceEmptySalts(configContent);
-
-      // Write the updated content back to wp-config.php
-      fs.writeFileSync(configPath, configContent, 'utf8');
-
-      console.log('🆗 WordPress configuration set up successfully.');
-    } catch (error) {
-      console.error('🔴 Error setting up WordPress configuration:', error);
-    }
-  }
-
-  /**
-   * Installs WordPress and sets up the WordPress configuration.
-   *
-   * @returns {Promise} A Promise that resolves when the installation and configuration are complete.
-   */
-  async install () {
-    const { version, language } = this.config.wordpress;
-    await this.installWordPress(version, language);
-    await this.setupWordPressConfig();
-  }
-}
-
-class PluginPackage extends Package {
-  /**
-   * Asynchronously installs a plugin.
-   *
-   * @return {Promise<void>} A Promise that resolves when the plugin is installed successfully.
-   */
-  async install () {
-    return await this.installPackage(this.config, 'plugin').then(() => {
-      console.log(`🆗 Plugin ${this.config.name} installed successfully.`);
-    });
-  }
-}
-
-class ThemePackage extends Package {
-  /**
-   * Installs a package with the specified name and version as a theme.
-   *
-   * @return {Promise<void>} A promise that resolves once the package is installed.
-   */
-  async install () {
-    return await this.installPackage(this.config, 'theme').then(() => {
-      console.log(`🆗 Theme ${this.config.name} installed successfully.`);
-    });
-  }
-}
-
-module.exports = {
-  WordPressPackage,
-  PluginPackage,
-  ThemePackage
-};
+module.exports = Package;
